refactor(tui): extract settings menu projection
