+++ core/crates/omegon/src/tui/mod.rs
@@ -43,6 +43,7 @@ pub mod segment_detail;
 pub mod segments;
 pub mod selector;
 pub(crate) mod settings_menu;
+mod settings_menu_projection;
 mod slash_commands;
 pub mod spinner;
 pub mod splash;
@@ -149,6 +150,8 @@ use self::menu_surface::{ActiveMenu, MenuMode};
 use self::permission_lane::{format_permission_prompt, permission_response_for_key};
 use self::segments::{SegmentContent, SegmentExportMode, SegmentRenderMode};
 use self::settings_menu::SelectorKind;
+#[cfg(test)]
+use self::settings_menu_projection::settings_profile_source_line;
 use self::slash_commands::SlashResult;
 use self::tutorial_state::TutorialState;
 #[cfg(test)]
@@ -655,18 +658,6 @@ fn editor_height_for(editor: &Editor, main_area: Rect) -> u16 {
     let editor_rows = editor.visual_line_count(content_width) as u16;
     let max_editor = (main_area.height * 40 / 100).clamp(5, 20);
     (editor_rows + 2).clamp(3, max_editor) // +2 for border
-}
-
-fn settings_profile_source_line(source: &crate::settings::ProfileSource) -> String {
-    match source {
-        crate::settings::ProfileSource::Project(path) => {
-            format!("profile: project · file: {}", path.display())
-        }
-        crate::settings::ProfileSource::User(path) => {
-            format!("profile: user · file: {}", path.display())
-        }
-        crate::settings::ProfileSource::BuiltInDefault => "profile: built-in defaults".to_string(),
-    }
 }
 
 impl App {
@@ -3533,7 +3524,17 @@ impl App {
             loaded_profile.source,
             &settings,
         );
-        let source_line = settings_profile_source_line(&drift.source);
+        let source_line = match &drift.source {
+            crate::settings::ProfileSource::Project(path) => {
+                format!("profile: project · file: {}", path.display())
+            }
+            crate::settings::ProfileSource::User(path) => {
+                format!("profile: user · file: {}", path.display())
+            }
+            crate::settings::ProfileSource::BuiltInDefault => {
+                "profile: built-in defaults".to_string()
+            }
+        };
         let drift_value = if drift.changed_count > 0 {
             format!("Δ{}", drift.changed_count)
         } else {
@@ -3897,12 +3898,6 @@ impl App {
     }
 
     fn settings_menu_projection(&self) -> crate::surfaces::menu::MenuProjection {
-        use crate::surfaces::menu::{
-            MenuActionProjection, MenuBadgeProjection, MenuBadgeTone, MenuGroupProjection,
-            MenuProjection, MenuRowKind, MenuRowProjection, MenuTabProjection,
-        };
-        use crate::surfaces::settings::{SettingsEditorProjection, SettingsStatusProjection};
-
         let settings = self.settings_projection();
         let settings_snapshot = self.settings();
         let loaded_profile = crate::settings::Profile::load_with_source(self.cwd());
@@ -3912,216 +3907,9 @@ impl App {
                 loaded_profile.source,
                 &settings_snapshot,
             );
-        let profile_source_line = settings_profile_source_line(&profile_drift.source);
-        let drift_line = if profile_drift.changed_count > 0 {
-            format!(
-                "runtime drift: Δ{} · /profile save or /profile apply · {profile_source_line}",
-                profile_drift.changed_count
-            )
-        } else {
-            format!("runtime drift: clean · {profile_source_line}")
-        };
-        let mut menu = MenuProjection::new("settings", "Settings");
-        menu.summary = Some(format!(
-            "Universal configuration entrypoint for runtime and capability settings. Enter opens or edits the selected area.\n{drift_line}"
-        ));
-        menu.footer = Some("↑/↓ navigate · Tab switch tabs · / filter · Enter open/edit · s save · a apply · n save named · Esc close".into());
-        let configuration_rows = [
-            (
-                "runtime",
-                "Runtime",
-                "Edit runtime, workspace, and inference defaults here.",
-                "/settings runtime",
-            ),
-            (
-                "model",
-                "Model & inference",
-                "Select model routes, providers, grades, and routing policy.",
-                "/model",
-            ),
-            (
-                "auth",
-                "Authentication",
-                "Manage provider credentials, login state, and vault unlock.",
-                "/auth",
-            ),
-            (
-                "skills",
-                "Skills",
-                "Install, inspect, refresh, and remove operator skills.",
-                "/skills",
-            ),
-            (
-                "extensions",
-                "Extensions",
-                "Install, enable, disable, update, and inspect extensions.",
-                "/extension",
-            ),
-            (
-                "ui",
-                "UI & presentation",
-                "Configure Om, Active, Full, and individual interface surfaces.",
-                "/ui",
-            ),
-            (
-                "context",
-                "Context",
-                "Configure context class and manage context lifecycle.",
-                "/context",
-            ),
-            (
-                "memory",
-                "Memory",
-                "Inspect memory configuration and current memory state.",
-                "/memory",
-            ),
-            (
-                "profile",
-                "Profiles",
-                "Inspect, apply, and persist project or user defaults.",
-                "/profile",
-            ),
-            (
-                "secrets",
-                "Secrets",
-                "Manage named secrets and credential values.",
-                "/secrets",
-            ),
-            (
-                "sandbox",
-                "Sandbox & permissions",
-                "Configure child isolation and workspace access policy.",
-                "/sandbox",
-            ),
-            (
-                "updates",
-                "Updates",
-                "Configure the release channel and install available updates.",
-                "/update",
-            ),
-        ]
-        .into_iter()
-        .map(|(id, label, description, command)| MenuRowProjection {
-            id: format!("settings.area.{id}"),
-            label: label.into(),
-            description: description.into(),
-            value: Some(command.into()),
-            kind: MenuRowKind::Action,
-            badges: Vec::new(),
-            metadata: vec!["configuration area".into(), command.into()],
-            primary_action: Some(MenuActionProjection::command(
-                format!("settings.area.{id}.open"),
-                "Open",
-                command,
-            )),
-            actions: Vec::new(),
-            safety: None,
-            availability: None,
-        })
-        .collect();
-        let configuration_tab = MenuTabProjection {
-            id: "configuration".into(),
-            label: "Configuration".into(),
-            groups: vec![MenuGroupProjection {
-                id: "settings.configuration".into(),
-                label: "Configuration areas".into(),
-                description: Some(
-                    "Canonical entrances for every operator-configurable capability.".into(),
-                ),
-                rows: configuration_rows,
-            }],
-        };
-        menu.tabs
-            .extend(settings.tabs.into_iter().map(|tab| MenuTabProjection {
-                id: tab.id.clone(),
-                label: tab.label.clone(),
-                groups: vec![MenuGroupProjection {
-                    id: format!("settings.{}", tab.id),
-                    label: tab.label,
-                    description: None,
-                    rows: tab
-                        .rows
-                        .into_iter()
-                        .map(|row| {
-                            let tone = match row.status {
-                                SettingsStatusProjection::Normal => MenuBadgeTone::Neutral,
-                                SettingsStatusProjection::Warning => MenuBadgeTone::Warning,
-                                SettingsStatusProjection::Error => MenuBadgeTone::Danger,
-                                SettingsStatusProjection::Disabled => MenuBadgeTone::Info,
-                            };
-                            let editor = match row.editor {
-                                SettingsEditorProjection::Choice => "choice",
-                                SettingsEditorProjection::Toggle => "toggle",
-                                SettingsEditorProjection::Text => "text",
-                                SettingsEditorProjection::Number => "number",
-                                SettingsEditorProjection::Action => "action",
-                                SettingsEditorProjection::ReadOnly => "read only",
-                            };
-                            let mut metadata =
-                                vec![row.persistence.label().to_string(), editor.to_string()];
-                            if let Some(profile) = row.profile {
-                                metadata.push(format!("profile: {}", profile.profile_value));
-                            }
-                            let row_id = row.id.clone();
-                            MenuRowProjection {
-                                id: row.id,
-                                label: row.label,
-                                description: row.description,
-                                value: Some(row.value),
-                                kind: MenuRowKind::Object,
-                                badges: vec![MenuBadgeProjection {
-                                    label: format!("{:?}", row.status).to_lowercase(),
-                                    tone,
-                                }],
-                                metadata,
-                                primary_action: Some(MenuActionProjection::open_settings_row(
-                                    format!("settings.{row_id}.open"),
-                                    "Edit",
-                                    row_id,
-                                )),
-                                actions: Vec::new(),
-                                safety: None,
-                                availability: None,
-                            }
-                        })
-                        .collect(),
-                }],
-            }));
-        menu.tabs.push(configuration_tab);
-        menu.actions = vec![
-            {
-                let mut action =
-                    MenuActionProjection::command("settings.save", "Save profile", "/profile save");
-                action.key = Some("s".into());
-                action
-            },
-            {
-                let mut action = MenuActionProjection::command(
-                    "settings.apply",
-                    "Apply profile",
-                    "/profile apply",
-                );
-                action.key = Some("a".into());
-                action
-            },
-            {
-                let mut action = MenuActionProjection::prime_editor(
-                    "settings.save_named_user",
-                    "Save as named (user)",
-                    "/profile save --name ",
-                    "Type the profile name and press Enter — saved to ~/.omegon/profiles/<name>.json",
-                );
-                action.key = Some("n".into());
-                action
-            },
-            MenuActionProjection::prime_editor(
-                "settings.save_named_project",
-                "Save as named (project)",
-                "/profile save --name ",
-                "Type the profile name followed by ' --project' and press Enter — saved to .omegon/profiles/<name>.json",
-            ),
-        ];
-        menu
+        settings_menu_projection::build_settings_menu_projection(
+            settings_menu_projection::SettingsMenuInputs::new(settings, profile_drift),
+        )
     }
 
     fn open_skills_menu(&mut self) -> Result<(), String> {
